requirements.txt
PySide6
mysql-connector-python
sqlalchemy